Adds the conductivity analyzer class
